+++ packages/yourloops/test/integration/pages/profile/hcp-profile.spec.tsx
@@ -58,10 +58,6 @@ describe('Profile page for hcp', () => {
     hcpProfession: HcpProfession.diabeto
   }
   const settings: Settings = {
-    a1c: {
-      date: '2020-01-01',
-      value: '7.5'
-    },
     country: CountryCodes.France,
     units: { bg: UnitsType.MMOLL }
   }