+++ packages/yourloops/test/unit/common/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2022, Diabeloop
+ * Copyright (c) 2021-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -31,15 +31,6 @@ import { AuthenticatedUserMetadata } from '../../../lib/auth/models/enums/authen
 import { UserRoles } from '../../../lib/auth/models/enums/user-roles.enum'
 import { UnitsType } from '../../../lib/units/models/enums/units-type.enum'
 import { CountryCodes } from '../../../lib/auth/models/country.model'
-
-const getNewHcp = (): User => {
-  return new User({
-    email: 'john.doe@example.com',
-    email_verified: true,
-    sub: 'auth0|123456789',
-    [AuthenticatedUserMetadata.Roles]: [UserRoles.hcp]
-  })
-}
 
 const getHcp = (): User => {
   const email = 'john.doe@example.com'
@@ -78,7 +69,7 @@ const getPatient = (): User => {
     sub: 'auth0|a0a0a0b0',
     [AuthenticatedUserMetadata.Roles]: [UserRoles.patient]
   })
-  patient.settings = { a1c: { date: '2020-01-01', value: '7.5' }, country: CountryCodes.France }
+  patient.settings = { a1c: { rawdate: '2020-01-01', date: '2020-01-01', value: '7.5' }, country: CountryCodes.France }
   patient.profile = {
     email,
     firstName: 'Josephine',
@@ -102,4 +93,4 @@ const getPatient = (): User => {
 /**
  * Logged-in users for test, choose one suitable
  */
-export const loggedInUsers = { getHcp, getPatient, getCaregiver, getNewHcp }
+export const loggedInUsers = { getHcp, getPatient, getCaregiver }