+++ packages/yourloops/test/unit/pages/profile/profile-page-context.hook.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -65,7 +65,7 @@ describe('Profile page context hook', () => {
   }
   const settings: Settings = {
     a1c: {
-      date: '2020-01-01',
+      rawdate: '2020-01-01',
       value: '7.5'
     },
     country: CountryCodes.France,